SpreadsheetEngineContextTesting2 extracted SpreadsheetEngineContextTesting
